Use Navigator for views
