+++ packages/playground/website/src/components/site-view/style.module.css
@@ -1,0 +1,13 @@
+.playground-view {
+	flex-grow: 1;
+	header {
+		max-height: 100px;
+		overflow: hidden;
+		transition: max-height 0.2s ease-in-out;
+	}
+}
+.playground-view-manager-preview {
+	header {
+		max-height: 0;
+	}
+}
